refactored the backend so it is cleaner and more readable, to do that I implemented blueprints in flask and sectioned out the routes for each auth and salesmen. I also created a delete salesman route to delete a salesmen
