+++ src/components/Header/UserMenu.jsx
@@ -8,7 +8,7 @@ import httpClient from '../../httpClient';
 export default function UserMenu() {
 
    const logoutUser = async () => {
-      await httpClient.post("//127.0.0.1:8000/logout");
+      await httpClient.post("//127.0.0.1:8000/auth/logout");
       window.location.href = "/";
     };
