+++ src/components/Salesman.jsx
@@ -10,7 +10,7 @@ export default function Salesman({ setTitle }) {
   useEffect(() => {
     (async () => {
       try {
-        const resp = await httpClient.get("//127.0.0.1:8000/@salesmen");
+        const resp = await httpClient.get("//127.0.0.1:8000/salesmen/@salesmen");
         console.log(resp.data)
         setSalesmen(resp.data)
       } catch (error) {
